requirements.dep.txt
Django==4.2.14
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
drf-yasg
django-cors-headers
gunicorn
psycopg==3.2.1
psycopg-binary==3.2.1
python-decouple==3.8
django-recaptcha
Pillow==10.4.0
django-redis
django-ckeditor-5
requests
flake8-django
black